Fix source-level compiler warnings across the codebase
Resolve warnings surfaced by GCC, Clang, and MSVC in a build of the
current tree. CI's ccache masks most of these after the first build, so
they were only partially visible in the workflow logs; a from-scratch
build exposes the full set.

Core / headers:
- Machine.hpp: cast uint64_t diagnostic-trace arguments to unsigned long
  long so the %llu format matches on LP64 targets (-Wformat).
- ExtensionManifest, PresetConfig, BreakpointEntry, WatchpointEntry: give
  every member a default member initializer so partial aggregate/designated
  initialisation no longer trips -Wmissing-field-initializers. Fixes the
  warning at all construction sites at once.
- Saa5050.hpp: make the 12-bit-to-8-bit hold-graphics store an explicit
  cast (C4244). NOTE: m_last_graphics_data is uint8_t but data carries 12
  bits here; the truncation is pre-existing behaviour, not introduced by
  this change -- worth a separate look at teletext hold-graphics.
- ExtensionRpc.hpp: add a serialized() helper that turns a protobuf
  SerializeToString() result into an RpcStatus, and route every extension
  dispatcher through it so a serialization failure surfaces as an internal
  error instead of being silently discarded (nodiscard / -Wunused-result).

acorn-scsi/HardDiskImage.cpp:
- Check the fread() result when loading a DSC and fall back to a generated
  default on a short read, rather than ignoring the return (-Wunused-result).

Tests:
- test_via_timing.cpp: three program builders declared the address cursor
  as size_t; make them uint16_t to match the rest of the file and stop the
  size_t->uint16_t narrowing on every machine.write() (C4267).
- test_cursor.cpp: rename an inner loop variable that shadowed an outer one
  (C4456).
- Remove dead code and unused symbols: tick_socket(), the bp2_id that was
  set but never read, and the unused raw pointer; mark the documentary
  REG_TX_LAST constant and an unused month read [[maybe_unused]].
- test_econet_fileserver.cpp: drop a trailing backslash in a // comment that
  formed a spurious line continuation (-Wcomment).

+++ src/core/include/beebium/Machine.hpp
@@ -220,7 +220,7 @@ public:
                             auto* ula = state_.memory.tube_socket.tube_ula();
                             uint16_t pc = state_.memory.tube_socket.diag_parasite_pc();
                             fprintf(stderr, "[STRETCH-INFO+%llu] tube_ula offset=%u (write stretch), parasite_pc=0x%04X\n",
-                                    since,
+                                    static_cast<unsigned long long>(since),
                                     ula ? ula->pending_offset() : 99,
                                     pc);
                         }
@@ -366,7 +366,7 @@ public:
                 if (ticks_since == 0 || ticks_since == 1000 || ticks_since == 10000 ||
                     ticks_since == 100000 || ticks_since == 500000) {
                     fprintf(stderr, "[DUMP tick+%llu] nmi_pending=%d, ff=%d, irq=%d, dev_nmi=0x%02X, nmi_flags=0x%02X, in_nmi=%d\n",
-                            ticks_since, econet_nmi,
+                            static_cast<unsigned long long>(ticks_since), econet_nmi,
                             state_.memory.econet_socket.nmi_enable_ff() ? 1 : 0,
                             da->irq_output() ? 1 : 0,
                             state_.cpu.device_nmi_flags, state_.cpu.nmi_flags,
@@ -770,7 +770,7 @@ private:
                 if (ticks_since == 0 || ticks_since == 1000 || ticks_since == 10000 ||
                     ticks_since == 100000 || ticks_since == 400000) {
                     fprintf(stderr, "[STRETCH+%llu] nmi_pending=%d, ff=%d, irq=%d, dev_nmi=0x%02X, nmi_flags=0x%02X, rx_fifo_empty=%d\n",
-                            ticks_since, econet_nmi,
+                            static_cast<unsigned long long>(ticks_since), econet_nmi,
                             state_.memory.econet_socket.nmi_enable_ff() ? 1 : 0,
                             da2->irq_output() ? 1 : 0,
                             state_.cpu.device_nmi_flags, state_.cpu.nmi_flags,
